Refactored GameApplication a little bit so it is possible to change the run method to do some custom stuff while no other thread is running. Changed Editor implementation to make use of that. Also the timer works correctly now in the Editor as long as we only have one world.

+++ Tools/EditorEngineProcess/ViewContext.cpp
@@ -204,7 +204,7 @@ void ezViewContext::RenderPassEventHandler(const ezPickingRenderPass::Event& e)
 
     ezMat4 mProj, mView;
 
-    m_Camera.GetProjectionMatrix(GetEditorWindow().m_uiWidth / GetEditorWindow().m_uiHeight, mProj);
+    m_Camera.GetProjectionMatrix((float)GetEditorWindow().m_uiWidth / GetEditorWindow().m_uiHeight, mProj);
     m_Camera.GetViewMatrix(mView);
 
     if (mProj.IsNaN())
